+++ linux/ws_main.py
@@ -124,6 +124,7 @@ main_src = SrcGroup(dir='src',
             'bp_sim_tcp.cpp',
             'utl_mbuf.cpp',
              'inet_pton.cpp',
+             'trex_global.cpp',
              'bp_sim.cpp',
              'trex_platform.cpp',
              'bp_sim_stf.cpp',
