+++ linux_dpdk/ws_main.py
@@ -249,6 +249,7 @@ main_src = SrcGroup(dir='src',
              'utl_policer.cpp',
              'astf/astf_template_db.cpp',
              'stt_cp.cpp',
+             'trex_global.cpp',
              'bp_sim.cpp',
              'trex_platform.cpp',
              'utl_term_io.cpp',
